Move Blaze helpers
