+++ app/emoji/client/index.js
@@ -1,4 +1,3 @@
 export { EmojiPicker } from './lib/EmojiPicker';
-export { renderEmoji } from './lib/emojiRenderer';
 export { emoji } from '../lib/rocketchat';
 export { createEmojiMessageRenderer } from './emojiParser';
