+++ app/ui-message/client/popup/messagePopupEmoji.html
@@ -1,4 +1,4 @@
 <template name="messagePopupEmoji">
-	{{> renderEmoji _id}}
+	{{{renderEmoji _id}}}
 	{{_id}}
 </template>
