+++ app/ui-utils/client/index.js
@@ -13,7 +13,7 @@ export { mainReady } from './lib/mainReady';
 export { Layout } from './lib/Layout';
 export { IframeLogin, iframeLogin } from './lib/IframeLogin';
 export { fireGlobalEvent } from './lib/fireGlobalEvent';
-export { getAvatarAsPng } from './lib/avatar';
+export { getAvatarAsPng } from './lib/getAvatarAsPng';
 export { popout } from './lib/popout';
 export { messageProperties } from '../lib/MessageProperties';
 export { MessageTypes } from '../lib/MessageTypes';
