+++ app/ui-utils/client/lib/getAvatarAsPng.js
@@ -1,8 +1,4 @@
-import { Blaze } from 'meteor/blaze';
-
 import { getUserAvatarURL } from '../../../utils/lib/getUserAvatarURL';
-
-Blaze.registerHelper('avatarUrlFromUsername', getUserAvatarURL);
 
 export const getAvatarAsPng = function(username, cb) {
 	const image = new Image();
