+++ client/templateHelpers/avatarUrlFromUsername.ts
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+import { getUserAvatarURL } from '../../app/utils/lib/getUserAvatarURL';
+
+Template.registerHelper('avatarUrlFromUsername', getUserAvatarURL);
