+++ client/templateHelpers/emojiUrlFromName.ts
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+import { getEmojiUrlFromName } from '../../app/emoji-custom/client/lib/emojiCustom';
+
+Template.registerHelper('emojiUrlFromName', getEmojiUrlFromName);
