+++ client/templateHelpers/index.ts
@@ -1,1 +1,5 @@
 import './log';
+import './renderEmoji';
+import './avatarUrlFromUsername';
+import './emojiUrlFromName';
+import './nrrargs';
