+++ client/templateHelpers/nrrargs.ts
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+Template.registerHelper('nrrargs', <T extends any[]>(...args: T) => ({
+	_arguments: args,
+}));
