+++ client/templateHelpers/renderEmoji.ts
@@ -1,0 +1,5 @@
+import { Template } from 'meteor/templating';
+
+import { renderEmoji } from '../lib/renderEmoji';
+
+Template.registerHelper('renderEmoji', renderEmoji);
